requirements.txt
# Python dependencies for TMDB Content Analytics Pipeline

# Apache Airflow
apache-airflow==2.8.1
apache-airflow-providers-google==10.14.0

# Google Cloud Platform
google-cloud-bigquery==3.14.1
google-cloud-core==2.4.1
google-auth==2.26.2

# API and HTTP
requests==2.31.0

# Data processing
pandas==2.1.4
numpy==1.26.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing (optional)
pytest==7.4.3
pytest-mock==3.12.0

# Code quality (optional)
black==23.12.1
flake8==7.0.0
mypy==1.8.0
